requirements.txt
anthropic==0.26.1
Flask==3.0.3
numpy==1.26.4
openai=1.30.3
python-dotenv==1.0.1
schema==0.7.5
tenacity==8.2.3
voyageai==0.2.2
Werkzeug==3.0.3
chromadb
